feature: NetworkManager creates docker network
